web/requirements.txt
alembic==0.9.6
annoy==1.12.0
asn1crypto==0.24.0
Babel==2.5.1
blinker==1.4
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
dominate==2.3.1
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-FlatPages==0.6
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gitdb2==2.0.5
GitPython==2.1.11
guess-language-spirit==0.5.3
gunicorn==19.9.0
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.5
Pillow==5.2.0
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.3
PyMySQL==0.9.2
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2017.2
PyYAML==3.13
redis==2.10.6
requests==2.18.4
six==1.11.0
smmap2==2.0.5
SQLAlchemy==1.1.14
urllib3==1.22
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
